+++ tests/invalid/InvalidInstanceField.php
@@ -24,20 +24,19 @@
  * THE SOFTWARE.
  */
 
-namespace macrominds\enum;
+namespace macrominds\enum\invalid;
 
-/**
- * @method static \App\Enums\Salutation MRS()
- * @method static \App\Enums\Salutation MR()
- * @method static \App\Enums\Salutation MS()
- */
-class Salutation
+use macrominds\enum\Enumerations;
+
+class InvalidInstanceField
 {
     use Enumerations;
 
-    protected static $enums = [
-            'MRS' => 1,
-            'MR' => 2,
-            'MS' => 3
-        ];
+    /**
+    * invalid, because this is an instance field. Not a static property.
+    */
+    /*static*/ protected $enums = [
+        'ONE' => 1,
+        'TWO' => 2
+    ];
 }